fix: move utils blake
