+++ hermes/crates/cardano-chain-follower/src/lib.rs
@@ -21,7 +21,6 @@ mod point;
 mod snapshot_id;
 mod stats;
 mod utils;
-mod utility;
 mod witness;
 
 pub use chain_sync_config::ChainSyncConfig;
