+++ hermes/crates/cardano-chain-follower/src/utils.rs
@@ -1,5 +1,7 @@
 //! Simple general purpose utility functions.
 
+use blake2b_simd::{self, Params};
+
 /// Convert T to an i16. (saturate if out of range.)
 pub(crate) fn i16_from_saturating<T: TryInto<i16>>(value: T) -> i16 {
     match value.try_into() {
@@ -26,4 +28,21 @@ pub(crate) fn usize_from_saturating<
             Err(_) => usize::MAX,
         }
     }
+}
+
+/// Convert the given value to `blake2b_244` array.
+pub(crate) fn blake2b_244(value: &[u8]) -> anyhow::Result<[u8; 28]> {
+    let h = Params::new().hash_length(28).hash(value);
+    let b = h.as_bytes();
+    b.try_into()
+        .map_err(|_| anyhow::anyhow!("Invalid length of blake2b_244, expected 28 got {}", b.len()))
+}
+
+#[allow(dead_code)]
+/// Convert the given value to `blake2b_256` array.
+pub(crate) fn blake2b_256(value: &[u8]) -> anyhow::Result<[u8; 32]> {
+    let h = Params::new().hash_length(32).hash(value);
+    let b = h.as_bytes();
+    b.try_into()
+        .map_err(|_| anyhow::anyhow!("Invalid length of blake2b_256, expected 32 got {}", b.len()))
 }
